msm: krait-regualtor-pmic: Enforce strict checks
Some registers in the PMIC gangs are expected to be set in
a certain way in the boot loader. If these are not set introduce
code to print error message. Also add a flag to indicate the settings
were not correct at boot up which aids in debugging issues via ramdump
analysis.

The current code intended to BUG if the settings were bad but that
code path was not enabled since the function returned immediately.

Change-Id: I0a6c8aeb88463029061153fc64df3f391ae1f701
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>