msm: remote_spinlock: Remove incorrect use of smd_private.h
smd_private.h is an internal header file used to share private internal
details of the SMD driver to related components for the purposes of
debugging, segregating functionality, etc.  Nothing in smd_private.h
is intended to be exposed for client use, and including it in a file
should be a special exception, not a normal usecase.

The remote_spinlock driver includes smd_private.h, but does not need it.
Remove the inclusion of the header file.

Change-Id: I3505302ef81a72505d723c11f1e6a728cb12b72c
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>