msm: subsystem_restart: Remove incorrect use of smd_private.h
smd_private.h is an internal header file used to share private internal
details of the SMD driver to related components for the purposes of
debugging, segregating functionality, etc.  Nothing in smd_private.h
is intended to be exposed for client use, and including it in a file
should be a special exception, not a normal usecase.

The subsystem_restart driver is using smd_private.h as a means to access
the platfrom_driver public API.  Substitute the platfrom_driver public api
headers for the uses of smd_private.h.

Change-Id: I87c171378822cb3b8a019f2624bc336a448e1105
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/subsystem_restart.c