msm: smd_pkt: Support multiple unique smd channels with the same name
When the Linux SMD driver detects a new channel, it creates a platform
device using the name of the channel as the name of the device, and uses
the channel type as the instance id in the platform device. Platform
drivers can only list the name of the device that the driver supports,
and only one platform driver is allowed to be registered for a given
device name.

The smd_pkt device open is failing for the smd channels which has same
name and tries to register same platform drivers again.

Add platfrom drivers registration check to avoid the multiple
registration of same platform drivers.

Change-Id: I9c187158d66bf52b1941e0f499b1a58196995073
Signed-off-by: Arun Kumar Neelakantam <aneela@codeaurora.org>