gpu: ion: make startswith more truthful
Currently, the startswith function in Ion returns true even if the
prefix is longer than the string being tested. For example:

    startswith("abc", "abcd"); // returns true

This isn't really a problem with the way it's being used currently but
it should be fixed for clarity and possible future uses of the
function. Always return false if the prefix is longer than the string
being tested.

Change-Id: Iabd45307be30278aaed5edcbd6e23045e0cb0c6c
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>