msm: smp2p: Test SMEM availability before probe completes
SMP2P depends on the SMEM driver and requires it to be in a ready state
before SMP2P can provide service to clients.  As a result, SMP2P should
check the state of SMEM before declaring SMP2P is ready.

Check the state of SMEM in the SMP2P probe, and appropriately handle cases
where SMEM is not ready at the time of the check.

Change-Id: I47913a36e58b4a7c922d1133e58249deb4f1796a
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/smp2p.c