msm: krait-regulator: fix unnecessary calls to switch to LDO
When asked to set a LDO voltage the driver rounds it up to
the nearest step value. However, when checking if the requested voltage
is already set it doesn't account for the rounding up.

Fix this by rounding up the requested value, which avoids
unnecessary calls to set the LDO voltage and switch to LDO.

CRs-Fixed: 609879
Change-Id: Ia5a201721e39eece49dea3f61fbc585f116d5060
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>