msm: kgsl: Move the list of GPU cores to a separate header
The list of GPU cores is only going to get bigger in the future. Put
the list in its own header and break out the individual field
definitions to make it easier to read.  Remove some A2XX only fields
that got left over from the great purge of 2013.

Change-Id: Ic0dedbadb3bd363c8733521b7146a671dc0225fa
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>

Conflicts:
	drivers/gpu/msm/adreno.c