msm: kgsl: Clean up OCMEM/GMEM
Consolidate and clean up how we treat OCMEM vs GMEM.  For all intents
and purposes they are mostly the same thing to us.  Use the gmem_ prefix
for the size and base and define a feature flag to indicate which cores
use OCMEM to avoid long lists of adreno_is_ conditionals.

Change-Id: Ic0dedbad30148a33474dcdeff67e6c0b0dba2fa7
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>

Conflicts:
	drivers/gpu/msm/adreno-gpulist.h
	drivers/gpu/msm/adreno.c
	drivers/gpu/msm/adreno.h
	drivers/gpu/msm/adreno_a3xx.c
	drivers/gpu/msm/adreno_a4xx.c