msm: kgsl: Make the dma_set_coherent_mask call universal
Now that we are required to send a non-NULL device ptr to
dma_alloc_coherent(), we need to interrogate the kernel to see if
the DMA controller can properly support the device's addressing
limitation. Therefore, moving the dma_set_coherent_mask to
adreno_probe before we allocate the ringbuffer.

Change-Id: I29e9c77e7551bc3447877aa5570e4d89fbb94a2e
Signed-off-by: Shrenuj Bansal <shrenujb@codeaurora.org>

Conflicts:
	drivers/gpu/msm/kgsl.c