msm: kgsl: Reset the bus_mod value when power is re-enabled
kgsl_pwrctrl_pwrlevel_change() resets the bus_mod level when the
power level is changed but the power level doesn't get changed
on re-init if constraints are enabled with the net effect of
turning on the bus at a lower level than we would expect.  Reset
bus_mod in  kgsl_pwrctrl_enable() to make sure it always comes
up at the expected level.

Change-Id: Ic0dedbad303cc938d91583b3e3309d232590b975
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>