iommu: msm: Remove unneeded code
The V7S mapping code is checking whether a new mapping will overlap an
old one so that we don't try to map and then fail with a partial
mapping. However, since this functionality was added the mapping code
has been changed to unwind any mappings in case of error. Thus the
check_range function is not needed anymore.

Change-Id: Id5b06512550680da44f9499f1b5577ebc063a441
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>

Conflicts:
	drivers/iommu/msm_iommu_pagetable.c